Move line gizmos to its own module
