+++ crates/bevy_gizmos/src/lines/line_joints.wgsl
