+++ crates/bevy_gizmos/src/lines/pipeline_3d.rs
@@ -1,9 +1,11 @@
 use crate::{
     config::{GizmoLineJoint, GizmoLineStyle, GizmoMeshConfig},
+    GizmoRenderSystem,
+};
+use super::{
     line_gizmo_vertex_buffer_layouts, line_joint_gizmo_vertex_buffer_layouts, DrawLineGizmo,
-    DrawLineJointGizmo, GizmoRenderSystem, GpuLineGizmo, LineGizmo,
-    LineGizmoUniformBindgroupLayout, SetLineGizmoBindGroup, LINE_JOINT_SHADER_HANDLE,
-    LINE_SHADER_HANDLE,
+    DrawLineJointGizmo, GpuLineGizmo, LineGizmo,
+    LineGizmoUniformBindgroupLayout, SetLineGizmoBindGroup,
 };
 use bevy_app::{App, Plugin};
 use bevy_asset::Handle;
@@ -32,6 +34,8 @@ use bevy_render::{
     Render, RenderApp, RenderSet,
 };
 use bevy_utils::tracing::error;
+
+use super::{LINE_JOINT_SHADER_HANDLE, LINE_SHADER_HANDLE};
 
 pub struct LineGizmo3dPlugin;
 impl Plugin for LineGizmo3dPlugin {
